Warning...
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@2796 111d7451-93fa-4e81-90d4-8039f76616e9

+++ bytesobj.c
@@ -858,7 +858,7 @@ static FAST_CALL MUST_CHECK Obj *iter_element(struct iter_s *v1, size_t i) {
         v1->iter = Obj(iter);
         iter->len = 1;
     }
-    iter->data[0] = (vv1->len < 0) ? ~vv1->data[i] : vv1->data[i];
+    iter->data[0] = (vv1->len < 0) ? (uint8_t)~vv1->data[i] : vv1->data[i];
     return Obj(iter);
 }
